+++ grunt.js
@@ -46,6 +46,7 @@ module.exports = function (grunt) {
           'src/pluginHooks.js',
           'src/plugins/contextMenu.js',
           'src/plugins/legacy.js',
+          'src/plugins/manualColumnResize.js',
 
           'src/3rdparty/jquery.autoresize.js',
           'src/3rdparty/sheetclip.js',