+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## [0.8.3](https://github.com/warpech/jquery-handsontable/tree/v0.8.3) (Jan 8, 2013)
 
-Feature:
+Features:
 - optimize jquery-autoresize plugin: arclones element is emptied before injecting another node
+- manual column resize (initial commit) [TODO: write tests]
 
 ## [0.8.2](https://github.com/warpech/jquery-handsontable/tree/v0.8.2) (Jan 7, 2013)